requirements.txt
numpy
torch
ffmpeg-python
scipy
tqdm